// src/components/product-card/plans-interface.tsx
interface PlanCardInterface {
    planName: string;
    name: string;
    data: number;
    texts: number;
    minutes: number;
    price: number;
    productCTA: string;
    rearImage: string;
    frontImage: string;
    rearImageAlt: string;
    frontImageAlt: string;
}

export default PlanCardInterface;